PacketStream: Properly detect disconnects in nonblocking reads.
We see a bug where it's possible to get net-ssh into an infite loop
spinning CPU after a connection has been closed. I am unable to
reliably reproduce it, so I can't be positive this is the fix, but I
have not seen a recurrence in our environment since adding this patch.