connection/channel+session: lower max packet size & allow manual size.
Per section 6.1 of RFC 4253 an SSH implementation MUST support a packet
size of 32768 bytes, however, any size above this is not mandatory.
Currently, the max packet size is 64Kb and can cause a channel open
failure for hosts which only support the required 32Kb packet size.

As a result, we now default to the required size of 32768 (0x8000) in
order to avoid any errors with such hosts.

Additionally, the user can now manually define the maximum packet and
window size, should they choose to in the event that they wish to
obtain performance gains from a larger packet/window size.