Use minimum timeout when calling IO#select
If keepalive is enabled and the wait parameter is given when starting
the event loop, we should pass the minimum of the two values through to
IO#select. Otherwise the event loop could block for too long, and
keepalive packets would not be sent as often as they should be.