Resolve relative paths in the Include directive properly
According to ssh_config(5), relative paths in the Include directive is
relative from the directory of the root file, not from that of the
including file.  If `~/.ssh/config` includes `~/.ssh/dir1/file1` which
has a line `Include dir2/file2`, ssh tries to include
`~/.ssh/dir2/file2`, not `~/.ssh/dir1/dir2/file2`.