Ensure functionality in FIPS enabled systems
FIPS enabled system are restricted in which algorithms may be used.
Without restricting the underlying cryptographic methods used by
net/ssh, applications built around the library may crash in unexpected
ways during use.

This patch detects whether or not non-FIPS algorithms can be used and
adjusts the default settiings accordingly.

Additionally transport/algorithms was updated to print the client and
server algorithms that are in use so that users can easily detect a
server/client algorithm mismatch in the future.

Closes #574