New options for :verify_host_key
Deprecate some existing values, replacing them with a set of words
that match the classes in the `Net::SSH::Verifiers` module. Values
would be replaced as follows:

- `false` becomes `:never`,
- `true` becomes `:accept_new_or_local_tunnel`,
- `:very` becomes `:accept_new`, and
- `:secure` becomes `:always`.

This is a nice improvement, improving data type consistency (they're
all symbols) and expressiveness (they reveal the underlying classes).

This change was preliminarily approved in
https://github.com/net-ssh/net-ssh/issues/532