Fix host key checking
The known_hosts file may contain keys associated with a hostname, an
ip-address, or both.

When validating a key, the net-ssh gem ensure that both the hostname and
the ip-address match beforce adding that key.  Thus, if the known_hosts
file only contains one of these two pieces of information, the host key
verification fails.

Instead of adding keys when both the hostname and the ip-address match,
add them when the user-supplied identification of the remote host match
an entry in the known_hosts file.  Optionaly, if `check_host_ip` is set
to true, the resolved IP address of the remote host is also checked.