Only check IP addresses against host_ip
A knwon_host file may contain any number of names or IP addresses valid
for a given node.  An entry may have multiple names but no IP address,
in this case we host_ip will never be found in hostlist.

Instead, first build the list of IP addresses present in hostlist.  Only
if this list is not empty, perfom ip host checking.