Add cached KnownHosts implementation
This adds a new class Net::SSH::KnownHosts::Cached which loads the
known_hosts files once, builds an index, and uses that when searching
for keys for a given host. An instance of this object can be passed
using the `known_hosts` option in order to avoid loading known_hosts
multiple times when making multiple connections to either the same host
or several hosts.

This behavior is implemented in a new class because it differs from the
existing implementation in the case where the content of the known_hosts
file changes outside the current process. With the standard
implementation, those changes will always be picked up the next time a
connection is made, whereas the Cached implementation won't
automatically incorporated changes made out-of-band.