Check for timeouts while reading headers during server version negotiating
Currently, if we connect to a non-SSH server, we'll potentially block forever
waiting to receive an SSH version string.  We'll read whatever data the server
sends into @header but if that data is, e.g., the banner message from an FTP
server.  If the server never sends a line that starts with "SSH-", we'll wait
around until the server breaks the connection.