Support SSH options and keys in Proxy::Jump
This adds support for SSH options on the command line with the -o flag.
You can pass them in as options when you instantiate the Proxy::Jump class.

Keys and key_data from the SSH connection are also passed to the ssh
proxy. The key_data strings are written to disk as Tempfiles and cleaned
up when the Proxy::Jump object is garbage collected to ensure they exist when
the ssh command is run.