Tell any ProxyCommand process to hang up.
If there's a pid in the IO, send it a SIGHUP on cleanup.

Solves a problem when using EC2 Instance Connect, which doesn't
handle shutdown across the tunnel the way a regular socket might,
leading to zombie sessions on the server and net-ssh hanging on
close.  May resolve other similar issues with unconventional proxy
configurations.

Found by comparison with OpenSSH, which also sends a HUP, so this
is behavioural parity.