Fix parsing of the algorithms from ssh_config
Based on ssh_config's documentation parsing of the algorithm configs work in the following way:

> Specifies the ciphers allowed and their order of preference.  Multiple ciphers must be comma-separated.  If the specified list begins with a ‘+’ character, then the specified
> ciphers will be appended to the default set instead of replacing them.  If the specified list begins with a ‘-’ character, then the specified ciphers (including wildcards) will
> be removed from the default set instead of replacing them.  If the specified list begins with a ‘^’ character, then the specified ciphers will be placed at the head of the
> default set.

This means only the very first character in the list is used to determine operation (append, prepend, remove), afterwards the rest of the config items will remain as-is, without any prefix

This change would modify how the config is parsed from the previous version, which expected each item to have it's own prefix to how ssh_config works. This was likely due to the wrong wording in older versions of the config's documentation that was fixed a while ago: https://github.com/openssh/openssh-portable/commit/4f9d75fbafde83d428e291516f8ce98e6b3a7c4b

More specifically this fixes an issue with recent version of MacOS (Sequoia 15 and above) where the default ssh config is the following:

```
Ciphers ^aes-128-gcm@openssh.com,aes256-gcm@openssh.conf
```

Current version of the code would drop the first item (as it doesn't understand the prepend operator), then add the second. If the first character would be a `+` then however the first item would have been added, with the second silently ignored.